requirements.txt
asgiref==3.11.0
Django==5.2.8
pillow==12.0.0
setuptools==80.9.0
sqlparse==0.5.3
tzdata==2025.2
wheel==0.45.1
djangorestframework==3.16.1